+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.9 - 2023-05-24
+### Fixed
+- Issue with Craft warning about initialising Twig before Craft
+- FractalTemplateLoader moved to its own class (PSR-1/2.3)
+
 ## 1.0.0 - 2018-02-21
 ### Added
 - Initial release